Create BIDS-Dependences-List.md
Create dependencies file list. As we discussed in issue #33 this is a working list of the software dependencies for new BIDS users. 
I included Octave in keeping with the open source theme.
Also worth noting that this file can be a precursor for the "requirements.txt" file that we will need for working with Jupyter/Binder.